+++ src/components/form/formStyles.js
@@ -1,0 +1,25 @@
+export const errorMessage = {
+  position: "absolute",
+  top: "110px",
+  color: "red",
+};
+export const formContainer = {
+  backgroundColor: "white",
+  height: "150px",
+  width: "400px",
+  borderRadius: "25px",
+  display: "flex",
+  flexDirection: "column",
+  justifyContent: "center",
+  alignItems: "center",
+  position: "relative",
+};
+export const formPageContainer = {
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  height: "600px",
+};
+export const siteTitle = {
+  textAlign: "center",
+};
