+++ src/context/weatherContext.js
@@ -3,10 +3,15 @@ import { createContext, useState } from "react";
 export const WeatherContext = createContext();
 
 export const WeatherContextProvider = ({ children }) => {
-  const [cityWeather, setCityWeather] = useState(null);
+  const [cityInput, setCityInput] = useState("");
 
   return (
-    <WeatherContext.Provider value={{ cityWeather, setCityWeather }}>
+    <WeatherContext.Provider
+      value={{
+        cityInput,
+        setCityInput,
+      }}
+    >
       {children}
     </WeatherContext.Provider>
   );
